requirements.txt
# Dependencies to run the app
Django==1.8.4
django-crispy-forms==1.4.0
django-lazysignup==1.0.1
# workaround lazy django-lazysignup dependencies
six==1.9.0

# Running on heroku
gunicorn==19.3.0
dj-database-url==0.3.0
psycopg2==2.6.1
dj-static==0.0.6

# Just for development
pyexpect==1.0.14
selenium==2.45.0
splinter==0.7.1
